# Conflicts:
#	koltinx-serialization-scale/build.gradle

+++ substrate-sdk-android/src/test/java/io/novasama/substrate_sdk_android/common/Math.kt
@@ -3,9 +3,9 @@ package io.novasama.substrate_sdk_android.common
 /**
  * Complexity: O(n * log(n))
  */
-fun List<Double>.median(): Double = sorted().let {
+fun List<Long>.median(): Double = sorted().let {
     val middleRight = it[it.size / 2]
-    val middleLeft = it[(it.size - 1) / 2] // will be same as middleRight if list size is odd
+    val middleLeft = it[(it.size - 1) / 2]
 
-    (middleLeft + middleRight) / 2
+    (middleLeft + middleRight) / 2.0
 }
